SQONE-462 Collapse page "content" into a single "route" controller & template partial.
